// src/api/interfaces/announcement.interface.ts
export interface announcementAttributes {
    Id: number
    AnnouncementTypeId: number
    AnnouncementName: string
    ShortDesc: string
    LongDesc: string
    AnnouncementDate: string
    ValidUntilDate: string
    ImageFile: string
    ExternalURL: string
    FlyerName: string
    DisplayOrder: number
    Active: boolean
}